services/bot/requirements.txt
alembic==1.8.1
asgiref==3.7.2
asyncpg==0.30.0
autopep8
black
dependency-injector==4.46.0
flake8
Flask-Login==0.6.2
nats-py==2.10.0
pydantic-settings==2.0.2
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0
python-dotenv==0.21.0
python-telegram-bot==22.6
pytz==2022.7.1
SQLAlchemy==2.0.19
taskiq==0.11.17
taskiq-dependencies==1.5.7
taskiq-nats==0.5.1
taskiq[reload]
Werkzeug==2.3.7
